remove duplicate SampleCodeTest.swift, add 1.x attachment code example.
